Refactor constants into shared library project (#64)
* Refactor SocialMediaMessage to shared Dapr.Test.Common.Models

* #57
Moved class to models folder

* #61
Move hashtag and constants to shared library

* Add docker-compose for CI

* Refactor SocialMediaMessage to shared Dapr.Test.Common.Models

* #57
Moved class to models folder

* #61
Move hashtag and constants to shared library

* Add docker-compose for CI

* Rebase from master

* Cleanup project dependencies

* Align docker-compose with ci build

* Adding non-ci github workflow

* Fix typo

* Downgraded docker-compose schema from 3.8 to 3.3

* fix ci validation-worker

* Add missing MIT licenses
