+++ common/Models/Constants.cs
@@ -1,0 +1,33 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+namespace Dapr.Tests.Common.Models
+{
+    public static class Constants
+    {
+        public static string[] HashTags = new string[]
+        {
+            "circle",
+            "ellipse",
+            "square",
+            "rectangle",
+            "triangle",
+            "star",
+            "cardioid",
+            "epicycloid",
+            "limocon",
+            "hypocycoid"
+        };
+
+        public static string[] Sentiments = new string[]
+        {
+            "verynegative",
+            "negative",
+            "neutral",
+            "positive",
+            "verypositive"
+        };
+    }
+}
