+++ docker-compose.yml
@@ -1,0 +1,38 @@
+version: '3.3'
+
+services:
+  feed-generator:
+    image: feed-generator
+    build:
+      context: .
+      dockerfile: feed-generator/Dockerfile
+      
+  hashtag-actor:
+    image: hashtag-actor
+    build:
+      context: .
+      dockerfile: hashtag-actor/Dockerfile
+
+  hashtag-counter:
+    image: hashtag-counter
+    build:
+      context: .
+      dockerfile: hashtag-counter/Dockerfile
+
+  message-analyzer:
+    image: message-analyzer
+    build:
+      context: .
+      dockerfile: message-analyzer/Dockerfile
+      
+  snapshot:
+    image: snapshot
+    build:
+      context: .
+      dockerfile: snapshot/Dockerfile
+
+  validation-worker:
+    image: validation-worker
+    build:
+      context: .
+      dockerfile: validation-worker/Dockerfile
