+++ hashtag-actor/hashtag-actor.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -9,6 +9,10 @@
     <PackageReference Include="Dapr.Actors" Version="0.7.0-preview01" />
     <PackageReference Include="Dapr.Actors.AspNetCore" Version="0.7.0-preview01" />
     <PackageReference Include="prometheus-net" Version="3.5.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\common\common.csproj" />
   </ItemGroup>
 
 </Project>
