+++ hashtag-counter/hashtag-counter.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="prometheus-net" Version="3.5.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\common\common.csproj" />
+  </ItemGroup>
+
 </Project>
