+++ message-analyzer/Startup.cs
@@ -6,6 +6,7 @@
 namespace MessageAnalyzer
 {
     using Dapr.Client;
+    using Dapr.Tests.Common.Models;
     using Google.Protobuf.WellKnownTypes;
     using Microsoft.AspNetCore.Builder;
     using Microsoft.AspNetCore.Hosting;
