+++ snapshot/snapshot.csproj
@@ -11,9 +11,12 @@
     <PackageReference Include="Dapr.Actors" Version="0.7.0-preview01" />
     <PackageReference Include="Dapr.Actors.AspNetCore" Version="0.7.0-preview01" />
     <PackageReference Include="Dapr.AspNetCore" Version="0.7.0-preview01" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.10.8" />
     <PackageReference Include="prometheus-net" Version="3.5.0" />
 
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\common\common.csproj" />
   </ItemGroup>
 
 </Project>
